Implement CSS

+++ React/foes-db/src/components/file-view-section/FileViewSection.jsx
@@ -1,5 +1,7 @@
+import { Button } from '@mui/material';
 import React from 'react';
 import { useSelector } from 'react-redux';
+import DownloadIcon from '@mui/icons-material/Download';
 
 const FileViewSection = (props) => {
 
@@ -10,14 +12,16 @@ const FileViewSection = (props) => {
   };
 
   const onDeleteHandler = () => {
-    fetch(`http://127.0.0.1:8000/api/deleteAttachment/${props.obj.id}`, {
-      method: 'DELETE',
-      headers: {
-        Authorization: `Bearer ${token}`
-      }
-    });
-    const newAttachments = (props.attachments).filter((element, i) => !(element.id === props.attachmentId));
-    props.setAttachments(newAttachments);
+    if (window.confirm("Are you sure you want to delete this element?") == true) {
+      fetch(`http://127.0.0.1:8000/api/deleteAttachment/${props.obj.id}`, {
+        method: 'DELETE',
+        headers: {
+          Authorization: `Bearer ${token}`
+        }
+      });
+      const newAttachments = (props.attachments).filter((element, i) => !(element.id === props.attachmentId));
+      props.setAttachments(newAttachments);
+    }
   };
 
   return (
@@ -32,10 +36,13 @@ const FileViewSection = (props) => {
         <div className="formInput">
           <label>file_name: {props.obj.file_name}</label>
         </div>
-        <a href={`http://127.0.0.1:8000/api/downloadAttachment/${props.attachmentId}`} download='myFile'>Download</a>
-        attachmentID: {props.attachmentId}
-        <button onClick={onEditHandler}>edit</button>
-        <button onClick={onDeleteHandler}>delete</button>
+        <a href={`http://127.0.0.1:8000/api/downloadAttachment/${props.attachmentId}`} download='myFile'>
+          <Button variant="contained" color='success' endIcon={<DownloadIcon />} onClick={() => { (<a href={`http://127.0.0.1:8000/api/downloadAttachment/${props.attachmentId}`} download='myFile'>Download</a>); }}>
+            Download
+          </Button>
+        </a>
+        <Button className='section-btn' variant='contained' onClick={onEditHandler}>edit</Button>
+        <Button className='section-btn' variant='outlined' color='error' onClick={onDeleteHandler}>delete</Button>
       </div>
     </div>
   );
